fix(meow) don't require predicate functions to be loaded
Custom meow states may require pulling in expensive code.
An exapmle of this is meow-lispy-mode which requires lispy
and lispy is an expensive package to load, which shouldn't
be done at the same time you want the modeline to be present.
Switching on meow--current-state (a symbol) doesn't require
eager loading of such packages